Fix multiple Win32 runtime and build issues found during code audit
Build fixes:
- Add trivia.o, auction.o, questmaster.o to Makefile, resolves all linker
  errors since the source files existed but were never compiled
- Fix strncasecmp implicit declaration in trivia.c with strings.h include

Critical Win32 runtime fix:
- Read loop in read_from_descriptor was gated on MSDOS or unix only,
  which excluded _WIN32 entirely. No client data was ever read on Windows.
  Add _WIN32 to the condition and use WSAGetLastError/WSAEWOULDBLOCK
  instead of errno/EWOULDBLOCK for Windows socket error handling.

getc EOF safety across all fread_* functions:
- fread_letter, fread_number, fread_to_eol, fread_word, fread_flag:
  change char c to int c and add EOF checks. getc returns int and
  storing in char loses EOF detection on platforms where char is unsigned.
- fread_to_eol: add EOF checks to prevent infinite loop at end of file.

Socket descriptor type:
- Change DESCRIPTOR_DATA.descriptor from int16_t to int. Socket handles
  on 64-bit Windows can exceed 32767, causing silent truncation.

+++ GodWars_Modern/src/merc.h
@@ -3411,7 +3411,7 @@ struct descriptor_data
     int32_t             outtop;         /* Current top of output buffer  */
     int32_t             outsize;        /* Allocated size of outbuf      */
     char               *host;
-    int16_t             descriptor;
+    int                 descriptor;
     int16_t             connected;
     bool                fcommand;
     char                inbuf[4 * MAX_INPUT_LENGTH];
